week 2.1
I meant to commit more changes in week 1, but I shalt try to start a habit of doing so now :). this commit is a bit late, and another will be made tomorrow when I complete the dynamic macro script. Came down with a cold (covid test negative), which is hitting me hard.