week 2.3
FINISHED THE spiral staircase. There was more work to be done to make it into a function and get it to work more seamlessly.

Now on to 3D modeling and rendering!

edit: i realized i didnt already commit a version of spiral staircase, so i guess this is the first one! Changed the name from week 2.2.2 to 2.3 because its a new assignment submission.