week 2.3.9
not yet done or anywhere with the 3d assignment, i got lost in a loop hole trying to figure out how to simulate water with a 30x30 plane using a noise map. I got pretty far, and chatGPT taught me a lot because of all the errors I went through.
Mel creates a 30x30 plane, a noise map, and tries to animate said plane using the noise map through the timeline.
Error is, when i try to access the noise map through the variable that holds said noise map, to set attributes so it can effect the height of the vertices, it cant recognize that the noise1 texture exists as an object in hypershader. ChatGPT doesn't know why either. So I'm going to stop here for tonight.

I got ahead of myself by wanted water for a sploosh sound for the rendered ball scene hahaha.